+++ src/Vector2D.h
@@ -12,7 +12,7 @@ public:
     Vector2D &Multiply(const Vector2D vector);
     Vector2D &Divide(const Vector2D vector);
 
-    Vector2D operator*(const int &i);
+    Vector2D operator*(const float &i);
 
     float xcomponent;
     float ycomponent;
